fixed workers type in production config

+++ production_server.py
@@ -9,7 +9,7 @@ import os
 logging.basicConfig(level=logging.INFO)
 
 # Calculate number of workers based on CPU cores
-workers = os.environ.get('NUM_WORKERS', (multiprocessing.cpu_count() * 2) + 1)
+workers = int(os.environ.get('NUM_WORKERS', (multiprocessing.cpu_count() * 2) + 1))
 print(f"Using {workers} workers")
 
 # Gunicorn config
